Libinject compiles and loads on Solaris. (but it does not work)
git-svn-id: http://ng/svn/edk/libinject/trunk@681 29bc9b9d-423a-0410-a51e-a9d8d852b884

+++ src/actions/dump.c
@@ -56,7 +56,7 @@ static bool ActionDump_perform(int pos, ActionData* data, SocketInfo* si,
   if (file == NULL) {
     char fname[FILENAME_MAX + 1];
     fname[FILENAME_MAX] = '\0';
-    if (snprintf(fname, FILENAME_MAX, "%s.%d", data[1].str, getpid()) > 0) {
+    if (snprintf(fname, FILENAME_MAX, "%s.%d", data[1].str, (int)getpid()) > 0) {
       data[0].p = file = fopen(fname, "w");
     }
     if (data[0].p == NULL) {
